*: Move list functions out of internal
